Refactoring into separate files.
Not necessary, but organizing might make things easier.
