+++ trapp/pages/pb/Components/FinishTurnButton.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const FinishTurnButton = (props) => {
+  return (
+    <button className="pbFinishTurn" onClick={props.onClick}>
+      Finish Turn
+    </button>
+  );
+};
