+++ trapp/pages/pb/Components/RackTile.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const RackTile = (props) => {
+  return (
+    <span className={props.tileclass} onClick={props.onClick}>
+      {props.tilevalue}
+    </span>
+  );
+};
