+++ trapp/pages/pb/Components/TileRecallButton.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const TileRecallButton = (props) => {
+  return (
+    <button className="pbRecallTiles" onClick={props.onClick}>
+      Recall Tiles
+    </button>
+  );
+};
