+++ trapp/pages/pb/constants/escapehatches.js
@@ -1,0 +1,11 @@
+export const escapehatches = [
+  "0-0",
+  "0-7",
+  "0-14",
+  "7-0",
+  "7-14",
+  "14-0",
+  "14-7",
+  "14-14",
+]; // coords of escape hatches
+
